sftp_to_blob_copy_az_function/requirements.txt
azure-functions
azure-durable-functions
azure-storage-blob
smbprotocol